Nuevo

+++ TransporteVehiculos/src/main/java/TransporteDao/DaoVehiculo.java
@@ -2,11 +2,11 @@ package TransporteDao;
 
 public interface DaoVehiculo {
 
-public double calcularDistancia(float velocidad);
+public float calcularDistancia(float velocidad);
     
-public double calcularDistancia(float velocidad, float tiempo);
+public float calcularDistancia(float velocidad, float tiempo);
      
-public double calcularDistancia(float velocidad, float tiempo, float eficiencia);
+public float calcularDistancia(float velocidad, float tiempo, float eficiencia);
       
 public abstract void transportar();
 }
